auth/requirements.txt
orjson==3.9.10

uvicorn==0.24.0.post1
uvloop==0.19.0

gunicorn==21.2.0

redis==5.0.1
asyncpg==0.29.0

opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0

fastapi==0.104.1
pydantic==2.4.2
python-multipart==0.0.6

fastapi-pagination==0.12.14

pydantic-settings==2.0.3

Authlib==1.3.0

sqlalchemy[asyncio]==2.0.23
sqlakeyset==2.0.1701229837
alembic==1.13.1

argon2-cffi==23.1.0

email-validator==2.1.0
httpx==0.26.0

Faker==23.1.0

sentry-sdk==1.42.0
